refactor: 음식 최종 선택 방식 수정
- 음식 최종 선택 방식을 여러 음식으로 받을 수 있도록 수정
- 여러 음식을 나타내므로 패키지 이름을 food에서 meal로 변경 (단, 응답 dto는 각각의 음식을 전달하므로 유지)
