--- projects/paygo-be/src/main/java/mju/paygo/food/domain/FoodRepository.java
@@ -1,9 +1,0 @@
-package mju.paygo.food.domain;
-
-import java.util.Optional;
-
-public interface FoodRepository {
-
-    Food save(Food food);
-    Optional<Food> findById(Long foodId);
-}
