--- projects/paygo-be/src/main/java/mju/paygo/food/domain/S3Uploader.java
@@ -1,9 +1,0 @@
-package mju.paygo.food.domain;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public interface S3Uploader {
-
-    String outerUpload(MultipartFile file, Long userId);
-    void deleteS3Object(String url);
-}
