--- projects/paygo-be/src/main/java/mju/paygo/food/exception/exceptions/FileConvertException.java
@@ -1,8 +1,0 @@
-package mju.paygo.food.exception.exceptions;
-
-public class FileConvertException extends RuntimeException {
-
-    public FileConvertException() {
-        super("파일 전환에 실패했습니다.");
-    }
-}
