--- projects/paygo-be/src/main/java/mju/paygo/food/exception/exceptions/FileTransferIOException.java
@@ -1,8 +1,0 @@
-package mju.paygo.food.exception.exceptions;
-
-public class FileTransferIOException extends RuntimeException {
-
-    public FileTransferIOException() {
-        super("파일 전송 과정에서 입출력 예외가 발생하였습니다.");
-    }
-}
