--- projects/paygo-be/src/main/java/mju/paygo/food/infrastructure/FoodJpaRepository.java
@@ -1,7 +1,0 @@
-package mju.paygo.food.infrastructure;
-
-import mju.paygo.food.domain.Food;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FoodJpaRepository extends JpaRepository<Food, Long> {
-}
