--- projects/paygo-be/src/main/java/mju/paygo/food/infrastructure/dto/PossibleCandidateFoodResponse.java
@@ -1,18 +1,0 @@
-package mju.paygo.food.infrastructure.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import java.math.BigDecimal;
-import java.util.List;
-
-public record PossibleCandidateFoodResponse(
-        @JsonProperty("keyname")
-        String keyName,
-
-        @JsonProperty("foodlist")
-        List<FoodInfoResponse> foods,
-
-        @JsonProperty("possibility")
-        BigDecimal possibility
-) {
-}
