+++ projects/paygo-be/src/main/java/mju/paygo/meal/domain/MealRepository.java
@@ -1,0 +1,9 @@
+package mju.paygo.meal.domain;
+
+import java.util.Optional;
+
+public interface MealRepository {
+
+    Meal save(Meal meal);
+    Optional<Meal> findById(Long foodId);
+}
