+++ projects/paygo-be/src/main/java/mju/paygo/meal/domain/S3Uploader.java
@@ -1,0 +1,9 @@
+package mju.paygo.meal.domain;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface S3Uploader {
+
+    String outerUpload(MultipartFile file, Long userId);
+    void deleteS3Object(String url);
+}
