+++ projects/paygo-be/src/main/java/mju/paygo/meal/exception/exceptions/FileConvertException.java
@@ -1,0 +1,8 @@
+package mju.paygo.meal.exception.exceptions;
+
+public class FileConvertException extends RuntimeException {
+
+    public FileConvertException() {
+        super("파일 전환에 실패했습니다.");
+    }
+}
