+++ projects/paygo-be/src/main/java/mju/paygo/meal/exception/exceptions/FileExtensionExtractException.java
@@ -1,0 +1,8 @@
+package mju.paygo.meal.exception.exceptions;
+
+public class FileExtensionExtractException extends RuntimeException {
+
+    public FileExtensionExtractException() {
+        super("파일 확장자 추출을 하지 못했습니다.");
+    }
+}
