+++ projects/paygo-be/src/main/java/mju/paygo/meal/infrastructure/MealJpaRepository.java
@@ -1,0 +1,7 @@
+package mju.paygo.meal.infrastructure;
+
+import mju.paygo.meal.domain.Meal;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MealJpaRepository extends JpaRepository<Meal, Long> {
+}
