+++ projects/paygo-be/src/main/java/mju/paygo/meal/infrastructure/dto/FoodAnalyzeResponse.java
@@ -1,0 +1,16 @@
+package mju.paygo.meal.infrastructure.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.List;
+
+public record FoodAnalyzeResponse(
+
+        String fileUrl,
+        @JsonProperty("predictlist")
+        List<PossibleCandidateFoodResponse> possibles
+) {
+        public FoodAnalyzeResponse withUrl(final String url) {
+                return new FoodAnalyzeResponse(url, possibles);
+        }
+}
