+++ projects/paygo-be/src/main/java/mju/paygo/meal/infrastructure/dto/FoodInfoResponse.java
@@ -1,0 +1,18 @@
+package mju.paygo.meal.infrastructure.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record FoodInfoResponse(
+        @JsonProperty("id")
+        Long id,
+
+        @JsonProperty("foodname")
+        String foodName,
+
+        @JsonProperty("manufacturer")
+        String manufacturer,
+
+        @JsonProperty("predict_key")
+        String predictKey
+) {
+}
